requirements.txt
pillow
tqdm
matplotlib
opencv-python
numpy
torchmetrics
torch-fidelity
scikit-image
bing-image-downloader